+++ R/sf_helpers.R
@@ -9,8 +9,8 @@
 #'
 #' @examples
 #' library(sf)
-#' snnu = read_sf(system.file('extdata/snnu.geojson',package = 'sdsfun'))
-#' sf_geometry_name(snnu)
+#' gzma = read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
+#' sf_geometry_name(gzma)
 #'
 sf_geometry_name = \(sfj){
   gname = attr(sfj, "sf_column")
@@ -28,8 +28,8 @@ sf_geometry_name = \(sfj){
 #'
 #' @examples
 #' library(sf)
-#' snnu = read_sf(system.file('extdata/snnu.geojson',package = 'sdsfun'))
-#' sf_geometry_type(snnu)
+#' gzma = read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
+#' sf_geometry_type(gzma)
 #'
 sf_geometry_type = \(sfj){
   sfj_type = sfj %>%
@@ -138,16 +138,17 @@ sf_distance_matrix = \(sfj){
 #'
 #' @examples
 #' library(sf)
-#' snnu = read_sf(system.file('extdata/snnu.geojson',package = 'sdsfun'))
-#' sf_utm_proj_wgs84(snnu)
+#' gzma = read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
+#' sf_utm_proj_wgs84(gzma)
 #'
 sf_utm_proj_wgs84 = \(sfj){
   if (!inherits(sfj,'sf')){
     sfj = sf::st_as_sf(sfj)
   }
-  longlat = dplyr::if_else(sf::st_is_longlat(sfj),TRUE,FALSE,FALSE)
-  if (!longlat){
-    stop("The spatial reference of the input `sfj` object needs to be in the WGS84 geographic coordinate system.")
+  crs_info = sf::st_crs(sfj)
+  iswgs84 = dplyr::if_else(crs_info$epsg == 4326,TRUE,FALSE,FALSE)
+  if (!iswgs84){
+    stop("The spatial reference of the input `sfj` object needs to be the WGS84 geographic coordinate system.")
   } else {
     sf_ext = as.double(sf::st_bbox(sfj))
     center_lon = mean(sf_ext[c(1,3)])