+++ NEWS.md
@@ -1,3 +1,17 @@
+# sdsfun 0.3.0
+
+* Begin to use `Rcpp` to complete some tasks with high computational complexity.
+
+* Add `dummy_vector` and `dummy_tbl` functions for dummy variables generation.
+
+* Add `spdep_nb` function for neighbours list generation.
+
+* Add `spdep_skater` function for spatial constrained clustering.
+
+* Add `fuzzyoverlay` function for spatial fuzzy overlay.
+
+* Update package extdata; now `sdsfun` has two extdata: `gzma.gpkg` and `pts.gpkg`.
+
 # sdsfun 0.2.1
 
 * Enhance support for other types of spatial vector data.
@@ -12,8 +26,7 @@
 
 # sdsfun 0.1.1
 
-* Update the function documentation for `spdep_contiguity_swm()` and 
-  `spdep_distance_swm()`.
+* Update the function documentation for `spdep_contiguity_swm()` and `spdep_distance_swm()`.
 
 * Move the source code from `SpatLyu/sdsfun` on GitHub to `stscl/sdsfun`.